# hlss_openapi.yaml
openapi: 3.1.0

info:
  title: High Level Screen Service (HLSS) API
  version: 0.1.0
  description: >
    HLSS provides application-specific logic and server-side rendered
    user interfaces for e-Ink devices, mediated by the Low Level Screen
    Service (LLSS). HLSS never communicates directly with devices.

servers:
  - url: https://lichess-hlss.example/api

security:
  - LLSSAuth: []

tags:
  - name: Instances
    description: Instance lifecycle and configuration
  - name: Inputs
    description: Input events forwarded by LLSS
  - name: Rendering
    description: Frame generation and rendering control

paths:

  /instances/init:
    post:
      tags: [Instances]
      summary: Initialize a new HLSS instance
      description: >
        Called by LLSS when a new instance is created.
        Establishes trust, stores callbacks, and initializes state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstanceInitRequest"
      responses:
        "200":
          description: Instance initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceInitResponse"

  /instances/{instance_id}/inputs:
    post:
      tags: [Inputs]
      summary: Receive input event from LLSS
      description: >
        Receives abstract button events forwarded by LLSS.
        HLSS updates internal state and may render a new frame.
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InputEvent"
      responses:
        "200":
          description: Input processed

  /instances/{instance_id}/status:
    get:
      tags: [Instances]
      summary: Get instance status
      description: >
        Returns the current status of the instance, including whether
        user configuration is required.
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Instance status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceStatus"

  /instances/{instance_id}/render:
    post:
      tags: [Rendering]
      summary: Force frame rendering
      description: >
        Optional endpoint allowing LLSS to request a fresh render
        (e.g. after reconnect or cache loss).
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Render scheduled

components:

  securitySchemes:
    LLSSAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    InstanceInitRequest:
      type: object
      required:
        - instance_id
        - callbacks
        - display
      properties:
        instance_id:
          type: string
        callbacks:
          type: object
          required: [frames, inputs, notify]
          properties:
            frames:
              type: string
              format: uri
            inputs:
              type: string
              format: uri
            notify:
              type: string
              format: uri
        display:
          type: object
          required: [width, height, bit_depth]
          properties:
            width:
              type: integer
            height:
              type: integer
            bit_depth:
              type: integer

    InstanceInitResponse:
      type: object
      required: [status, needs_configuration]
      properties:
        status:
          type: string
          enum: [initialized]
        needs_configuration:
          type: boolean
        configuration_url:
          type: string
          format: uri
          nullable: true

    InstanceStatus:
      type: object
      properties:
        instance_id:
          type: string
        ready:
          type: boolean
        needs_configuration:
          type: boolean
        configuration_url:
          type: string
          format: uri
          nullable: true
        active_screen:
          type: string
          description: >
            Logical screen currently active (e.g. new_match, game_123)

    InputEvent:
      type: object
      required: [button, event_type, timestamp]
      properties:
        button:
          type: string
          enum:
            - BTN_1
            - BTN_2
            - BTN_3
            - BTN_4
            - BTN_5
            - BTN_6
            - BTN_7
            - BTN_8
            - ENTER
            - ESC
            - HL_LEFT
            - HL_RIGHT
        event_type:
          type: string
          enum: [PRESS, LONG_PRESS, RELEASE]
        timestamp:
          type: string
          format: date-time